+++ src/eva2/tools/JarResources.java
@@ -96,11 +96,11 @@ public final class JarResources {
                 if (size == -1) {
                     size = ((Integer) htSizes.get(ze.getName())).intValue();
                 }
-                byte[] b = new byte[(int) size];
+                byte[] b = new byte[size];
                 int rb = 0;
                 int chunk = 0;
-                while (((int) size - rb) > 0) {
-                    chunk = zis.read(b, rb, (int) size - rb);
+                while ((size - rb) > 0) {
+                    chunk = zis.read(b, rb, size - rb);
                     if (chunk == -1) {
                         break;
                     }
